--- src/tic-tac-toe.h
@@ -1,8 +1,0 @@
-﻿// tic-tac-toe.h : включаемый файл для стандартных системных включаемых файлов
-// или включаемые файлы для конкретного проекта.
-
-#pragma once
-
-#include <iostream>
-
-// TODO: установите здесь ссылки на дополнительные заголовки, требующиеся для программы.
